requirements.txt
pylint
coverage
pytest
pytest-cov
pytest-html
pytest-watch
black
ruamel.yaml
requests
humanize
Faker
